+++ src/main/java/sapienza/iot/Interfaces/ReadLocationSensor.java
@@ -1,7 +1,6 @@
 package sapienza.iot.Interfaces;
 
 public interface ReadLocationSensor {
-    Long getLongitude();
-
-    Long getLatitude();
+    Double getLongitude();
+    Double getLatitude();
 }
